IGAPP-261: Fix app crashing if firebase is stripped
