+++ native/src/App.tsx
@@ -1,8 +1,7 @@
 import NetInfo from '@react-native-community/netinfo'
-import messaging from '@react-native-firebase/messaging'
 import { LinkingOptions, NavigationContainer } from '@react-navigation/native'
 import React, { ReactElement, useCallback, useState } from 'react'
-import { Linking, LogBox } from 'react-native'
+import { LogBox } from 'react-native'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
 import { enableScreens } from 'react-native-screens'
 import { OverflowMenuProvider } from 'react-navigation-header-buttons'
@@ -10,7 +9,7 @@ import { Provider } from 'react-redux'
 import { Store } from 'redux'
 import { ThemeProvider } from 'styled-components'
 
-import { CLOSE_PAGE_SIGNAL_NAME, LOCAL_NEWS_TYPE, NEWS_ROUTE, REDIRECT_ROUTE, setUserAgent } from 'api-client'
+import { CLOSE_PAGE_SIGNAL_NAME, REDIRECT_ROUTE, setUserAgent } from 'api-client'
 
 import NavigatorContainer from './NavigatorContainer'
 import AppStateListener from './components/AppStateListener'
@@ -23,14 +22,12 @@ import { RoutesParamsType } from './constants/NavigationTypes'
 import buildConfig from './constants/buildConfig'
 import { userAgent } from './constants/endpoint'
 import useSendOfflineJpalSignals from './hooks/useSendOfflineJpalSignals'
-import urlFromRouteInformation from './navigation/url'
 import { StateType } from './redux/StateType'
 import { StoreActionType } from './redux/StoreActionType'
 import createReduxStore from './redux/createReduxStore'
-import appSettings from './utils/AppSettings'
 import { DataContainer } from './utils/DataContainer'
 import DefaultDataContainer from './utils/DefaultDataContainer'
-import { pushNotificationsEnabled } from './utils/PushNotificationsManager'
+import { initializePushNotificationListener } from './utils/PushNotificationsManager'
 import sendTrackingSignal from './utils/sendTrackingSignal'
 
 enableScreens(true)
@@ -58,35 +55,7 @@ const linking: LinkingOptions<RoutesParamsType> = {
       }
     ]
   }),
-  subscribe: pushNotificationsEnabled()
-    ? (listener: (url: string) => void) => {
-        const onReceiveURL = ({ url }: { url: string }) => listener(url)
-
-        const onReceiveURLListener = Linking.addListener('url', onReceiveURL)
-
-        // TODO IGAPP-263: Temporary workaround until cityCode, languageCode and newsId are part of the push notifications
-        const unsubscribeNotification = messaging().onNotificationOpenedApp(() => {
-          appSettings.loadSettings().then(settings => {
-            const { selectedCity, contentLanguage } = settings
-            if (selectedCity && contentLanguage) {
-              listener(
-                urlFromRouteInformation({
-                  cityCode: selectedCity,
-                  languageCode: contentLanguage,
-                  route: NEWS_ROUTE,
-                  newsType: LOCAL_NEWS_TYPE
-                })
-              )
-            }
-          })
-        })
-
-        return () => {
-          onReceiveURLListener.remove()
-          unsubscribeNotification()
-        }
-      }
-    : undefined
+  subscribe: initializePushNotificationListener
 }
 const dataContainer: DataContainer = new DefaultDataContainer()
 const store: Store<StateType, StoreActionType> = createReduxStore(dataContainer)
