+++ native/src/utils/PushNotificationsManager.ts
@@ -1,7 +1,16 @@
-import messaging from '@react-native-firebase/messaging'
+import { FirebaseMessagingTypes } from '@react-native-firebase/messaging'
+import { Linking } from 'react-native'
+
+import { LOCAL_NEWS_TYPE, NEWS_ROUTE } from 'api-client'
 
 import buildConfig from '../constants/buildConfig'
+import urlFromRouteInformation from '../navigation/url'
+import appSettings from './AppSettings'
 import { log, reportError } from './sentry'
+
+// @ts-ignore The type of the import is wrong
+const importFirebaseMessaging = async (): Promise<() => FirebaseMessagingTypes.Module> =>
+  import('@react-native-firebase/messaging')
 
 export const pushNotificationsEnabled = (): boolean =>
   buildConfig().featureFlags.pushNotifications && !buildConfig().featureFlags.floss
@@ -12,6 +21,7 @@ export const requestPushNotificationPermission = async (): Promise<boolean> => {
     return false
   }
 
+  const messaging = await importFirebaseMessaging()
   const authStatus = await messaging().requestPermission()
   log(`Authorization status: ${authStatus}`)
   // Firebase returns either 1 or 2 for granted or 0 for rejected permissions
@@ -29,6 +39,7 @@ export const unsubscribeNews = async (city: string, language: string): Promise<v
   const topic = newsTopic(city, language)
 
   try {
+    const messaging = await importFirebaseMessaging()
     await messaging().unsubscribeFromTopic(topic)
   } catch (e) {
     reportError(e)
@@ -44,9 +55,46 @@ export const subscribeNews = async (city: string, language: string): Promise<voi
   const topic = newsTopic(city, language)
 
   try {
+    const messaging = await importFirebaseMessaging()
     await messaging().subscribeToTopic(topic)
   } catch (e) {
     reportError(e)
   }
   log(`Subscribed to ${topic} topic!`)
+}
+
+export const initializePushNotificationListener = (listener: (url: string) => void): (() => void) | void => {
+  if (pushNotificationsEnabled()) {
+    importFirebaseMessaging()
+      .then(messaging => {
+        const onReceiveURL = ({ url }: { url: string }) => listener(url)
+
+        const onReceiveURLListener = Linking.addListener('url', onReceiveURL)
+
+        // TODO IGAPP-263: Temporary workaround until cityCode, languageCode and newsId are part of the push notifications
+        const unsubscribeNotification = messaging().onNotificationOpenedApp(() => {
+          appSettings.loadSettings().then(settings => {
+            const { selectedCity, contentLanguage } = settings
+            if (selectedCity && contentLanguage) {
+              listener(
+                urlFromRouteInformation({
+                  cityCode: selectedCity,
+                  languageCode: contentLanguage,
+                  route: NEWS_ROUTE,
+                  newsType: LOCAL_NEWS_TYPE
+                })
+              )
+            }
+          })
+        })
+
+        return () => {
+          onReceiveURLListener.remove()
+          unsubscribeNotification()
+        }
+      })
+      .catch(() => log('Failed to import firebase'))
+  }
+
+  return undefined
 }
